Reorganize early-stage string validation
This commit moves the unparsed full-string validation to stringParser
and combines the remaining common event- and string-level top-level
function code under the initial validation function.
