+++ validator/event.js
@@ -11,11 +11,6 @@ const { buildSchemaAttributesObject } = require('./schema')
 const { Schemas } = require('../utils/schema')
 const { convertHedStringToLong } = require('../converter/converter')
 
-const openingGroupCharacter = '('
-const closingGroupCharacter = ')'
-const comma = ','
-const delimiters = [comma]
-
 const uniqueType = 'unique'
 const requiredType = 'required'
 const requireChildType = 'requireChild'
@@ -27,134 +22,6 @@ const dateTimeUnitClass = 'dateTime'
 const timeUnitClass = 'time'
 
 // Validation tests
-
-/**
- * Substitute certain illegal characters and report warnings when found.
- */
-const substituteCharacters = function (hedString) {
-  const issues = []
-  const illegalCharacterMap = { '\0': ['ASCII NUL', ' '] }
-  const flaggedCharacters = /[^\w\d./$ :-]/g
-  const replaceFunction = function (match, offset) {
-    if (match in illegalCharacterMap) {
-      const [name, replacement] = illegalCharacterMap[match]
-      issues.push(
-        generateIssue('invalidCharacter', {
-          character: name,
-          index: offset,
-          string: hedString,
-        }),
-      )
-      return replacement
-    } else {
-      return match
-    }
-  }
-  const fixedString = hedString.replace(flaggedCharacters, replaceFunction)
-
-  return [fixedString, issues]
-}
-
-/**
- * Check if group parentheses match. Pushes an issue if they don't match.
- */
-const countTagGroupParentheses = function (hedString) {
-  const issues = []
-  const numberOfOpeningParentheses = utils.string.getCharacterCount(
-    hedString,
-    openingGroupCharacter,
-  )
-  const numberOfClosingParentheses = utils.string.getCharacterCount(
-    hedString,
-    closingGroupCharacter,
-  )
-  if (numberOfOpeningParentheses !== numberOfClosingParentheses) {
-    issues.push(
-      generateIssue('parentheses', {
-        opening: numberOfOpeningParentheses,
-        closing: numberOfClosingParentheses,
-      }),
-    )
-  }
-  return issues
-}
-
-/**
- * Check if a comma is missing after an opening parenthesis.
- */
-const isCommaMissingAfterClosingParenthesis = function (
-  lastNonEmptyCharacter,
-  currentCharacter,
-) {
-  return (
-    lastNonEmptyCharacter === closingGroupCharacter &&
-    !(
-      delimiters.includes(currentCharacter) ||
-      currentCharacter === closingGroupCharacter
-    )
-  )
-}
-
-/**
- * Check for delimiter issues in a HED string (e.g. missing commas adjacent to groups, extra commas or tildes).
- */
-const findDelimiterIssuesInHedString = function (hedString) {
-  const issues = []
-  let lastNonEmptyValidCharacter = ''
-  let lastNonEmptyValidIndex = 0
-  let currentTag = ''
-  for (let i = 0; i < hedString.length; i++) {
-    const currentCharacter = hedString.charAt(i)
-    currentTag += currentCharacter
-    if (utils.string.stringIsEmpty(currentCharacter)) {
-      continue
-    }
-    if (delimiters.includes(currentCharacter)) {
-      if (currentTag.trim() === currentCharacter) {
-        issues.push(
-          generateIssue('extraDelimiter', {
-            character: currentCharacter,
-            index: i,
-            string: hedString,
-          }),
-        )
-        currentTag = ''
-        continue
-      }
-      currentTag = ''
-    } else if (currentCharacter === openingGroupCharacter) {
-      if (currentTag.trim() === openingGroupCharacter) {
-        currentTag = ''
-      } else {
-        issues.push(generateIssue('invalidTag', { tag: currentTag }))
-      }
-    } else if (
-      isCommaMissingAfterClosingParenthesis(
-        lastNonEmptyValidCharacter,
-        currentCharacter,
-      )
-    ) {
-      issues.push(
-        generateIssue('commaMissing', {
-          tag: currentTag.slice(0, -1),
-        }),
-      )
-      break
-    }
-    lastNonEmptyValidCharacter = currentCharacter
-    lastNonEmptyValidIndex = i
-  }
-  if (delimiters.includes(lastNonEmptyValidCharacter)) {
-    issues.push(
-      generateIssue('extraDelimiter', {
-        character: lastNonEmptyValidCharacter,
-        index: lastNonEmptyValidIndex,
-        string: hedString,
-      }),
-    )
-  }
-  return issues
-}
 
 const camelCase = /([A-Z-]+\s*[a-z-]*)+/
 
@@ -845,21 +712,6 @@ const checkForInvalidTopLevelTagGroupTags = function (
 // Validation groups
 
 /**
- * Validate the full unparsed HED string.
- *
- * @param {string} hedString The unparsed HED string.
- * @return {Issues[][]} String substitution issues and other issues.
- */
-const validateFullUnparsedHedString = function (hedString) {
-  const [fixedHedString, substitutionIssues] = substituteCharacters(hedString)
-  const issues = [].concat(
-    countTagGroupParentheses(fixedHedString),
-    findDelimiterIssuesInHedString(fixedHedString),
-  )
-  return [substitutionIssues, issues]
-}
-
-/**
  * Validate the full parsed HED string.
  *
  * @param {ParsedHedString} parsedString The parsed HED string to validate.
@@ -936,8 +788,7 @@ const validateIndividualHedTags = function (
 ) {
   let issues = []
   let previousTag = new ParsedHedTag('', '', [0, 0], hedSchemas)
-  for (let i = 0; i < parsedString.tags.length; i++) {
-    const tag = parsedString.tags[i]
+  for (const tag of parsedString.tags) {
     issues = issues.concat(
       validateIndividualHedTag(
         tag,
@@ -1084,61 +935,33 @@ const validateTopLevelTagGroups = function (
  *
  * @param {string|ParsedHedString} hedString The HED string to validate.
  * @param {Schemas} hedSchemas The HED schemas to validate against.
- * @param {boolean} doSemanticValidation Whether to perform semantic validation.
- * @return {[boolean, Issue[], ParsedHedString]} Whether validation passed, any issues, and any parsed string data.
+ * @return {[ParsedHedString, Schemas, Issue[], boolean]} The parsed HED string, the actual HED schema collection to use, any issues found, and whether to perform semantic validation.
  */
-const initiallyValidateHedString = function (
-  hedString,
-  hedSchemas,
-  doSemanticValidation,
-) {
-  if (doSemanticValidation) {
-    if (!hedSchemas.baseSchema.attributes) {
-      hedSchemas.baseSchema.attributes = buildSchemaAttributesObject(
-        hedSchemas.baseSchema.xmlData,
-      )
-    }
+const initiallyValidateHedString = function (hedString, hedSchemas) {
+  let doSemanticValidation = hedSchemas instanceof Schemas
+  if (!doSemanticValidation) {
+    hedSchemas = new Schemas(null)
   }
+  let parsedString, parsedStringIssues
   // Skip parsing if we're passed an already-parsed string.
   if (hedString instanceof ParsedHedString) {
-    const [substitutionIssues, fullHedStringIssues] =
-      validateFullUnparsedHedString(hedString.hedString)
-    if (fullHedStringIssues.length !== 0) {
-      return [
-        false,
-        false,
-        fullHedStringIssues.concat(substitutionIssues),
-        null,
-      ]
-    }
-    return [true, true, substitutionIssues, hedString]
+    parsedString = hedString
+    parsedStringIssues = []
   } else {
-    const [substitutionIssues, fullHedStringIssues] =
-      validateFullUnparsedHedString(hedString)
-    if (fullHedStringIssues.length !== 0) {
-      return [
-        false,
-        false,
-        fullHedStringIssues.concat(substitutionIssues),
-        null,
-      ]
-    }
-
-    const [parsedString, parsedStringIssues] = parseHedString(
-      hedString,
-      hedSchemas,
-    )
-    if (parsedStringIssues.length !== 0) {
-      return [
-        true,
-        false,
-        parsedStringIssues.concat(substitutionIssues),
-        parsedString,
-      ]
-    }
-
-    return [true, true, substitutionIssues, parsedString]
+    ;[parsedString, parsedStringIssues] = parseHedString(hedString, hedSchemas)
   }
+  if (parsedString === null) {
+    return [null, hedSchemas, parsedStringIssues, null]
+  } else if (parsedStringIssues.length > 0) {
+    doSemanticValidation = false
+    hedSchemas = new Schemas(null)
+  }
+  if (doSemanticValidation && !hedSchemas.baseSchema.attributes) {
+    hedSchemas.baseSchema.attributes = buildSchemaAttributesObject(
+      hedSchemas.baseSchema.xmlData,
+    )
+  }
+  return [parsedString, hedSchemas, parsedStringIssues, doSemanticValidation]
 }
 
 /**
@@ -1156,36 +979,30 @@ const validateHedString = function (
   checkForWarnings = false,
   expectValuePlaceholderString = false,
 ) {
-  let doSemanticValidation = hedSchemas instanceof Schemas
-  if (!doSemanticValidation) {
-    hedSchemas = new Schemas(null)
-  }
-  const [fullStringValid, parsedStringValid, initialIssues, parsedString] =
-    initiallyValidateHedString(hedString, hedSchemas, doSemanticValidation)
-  if (!fullStringValid) {
-    return [false, initialIssues]
-  } else if (!parsedStringValid) {
-    doSemanticValidation = false
-    hedSchemas = new Schemas(null)
+  const [
+    parsedString,
+    actualHedSchemas,
+    parsedStringIssues,
+    doSemanticValidation,
+  ] = initiallyValidateHedString(hedString, hedSchemas)
+  if (parsedString === null) {
+    return [false, parsedStringIssues]
   }
 
-  const issues = initialIssues.concat(
+  const issues = parsedStringIssues.concat(
     validateFullParsedHedString(parsedString, expectValuePlaceholderString),
     validateIndividualHedTags(
       parsedString,
-      hedSchemas,
+      actualHedSchemas,
       doSemanticValidation,
       checkForWarnings,
       false,
       expectValuePlaceholderString,
     ),
-    validateHedTagGroups(parsedString, hedSchemas, doSemanticValidation),
+    validateHedTagGroups(parsedString, actualHedSchemas, doSemanticValidation),
   )
-  if (issues.length === 0) {
-    return [true, []]
-  } else {
-    return [false, issues]
-  }
+
+  return [issues.length === 0, issues]
 }
 
 /**
@@ -1201,39 +1018,40 @@ const validateHedEvent = function (
   hedSchemas,
   checkForWarnings = false,
 ) {
-  const doSemanticValidation = hedSchemas instanceof Schemas
-  if (!doSemanticValidation) {
-    hedSchemas = new Schemas(null)
-  }
-  const [fullStringValid, parsedStringValid, initialIssues, parsedString] =
-    initiallyValidateHedString(hedString, hedSchemas, doSemanticValidation)
-  if (!(fullStringValid && parsedStringValid)) {
-    return [false, initialIssues]
+  const [
+    parsedString,
+    actualHedSchemas,
+    parsedStringIssues,
+    doSemanticValidation,
+  ] = initiallyValidateHedString(hedString, hedSchemas)
+  if (parsedString === null) {
+    return [false, parsedStringIssues]
   }
 
-  const issues = initialIssues.concat(
+  const issues = parsedStringIssues.concat(
     validateTopLevelTags(
       parsedString,
-      hedSchemas,
+      actualHedSchemas,
       doSemanticValidation,
       checkForWarnings,
     ),
-    validateTopLevelTagGroups(parsedString, hedSchemas, doSemanticValidation),
+    validateTopLevelTagGroups(
+      parsedString,
+      actualHedSchemas,
+      doSemanticValidation,
+    ),
     validateIndividualHedTags(
       parsedString,
-      hedSchemas,
+      actualHedSchemas,
       doSemanticValidation,
       checkForWarnings,
       true,
     ),
-    validateHedTagLevels(parsedString, hedSchemas, doSemanticValidation),
+    validateHedTagLevels(parsedString, actualHedSchemas, doSemanticValidation),
     validateHedTagGroups(parsedString),
   )
-  if (issues.length === 0) {
-    return [true, []]
-  } else {
-    return [false, issues]
-  }
+
+  return [issues.length === 0, issues]
 }
 
 module.exports = {
