+++ validator/stringParser.js
@@ -1,10 +1,13 @@
 const differenceWith = require('lodash/differenceWith')
+const zip = require('lodash/zip')
 
 const utils = require('../utils')
+const { generateIssue } = require('../utils/issues/issues')
 const { convertPartialHedStringToLong } = require('../converter/converter')
 
 const openingGroupCharacter = '('
 const closingGroupCharacter = ')'
+const delimiters = new Set([','])
 
 /**
  * A parsed HED substring.
@@ -298,7 +301,6 @@ const splitHedString = function (
   hedSchemas,
   groupStartingIndex = 0,
 ) {
-  const delimiter = ','
   const doubleQuoteCharacter = '"'
   const invalidCharacters = ['{', '}', '[', ']', '~']
 
@@ -324,7 +326,7 @@ const splitHedString = function (
     } else if (character === closingGroupCharacter) {
       groupDepth--
     }
-    if (groupDepth === 0 && character === delimiter) {
+    if (groupDepth === 0 && delimiters.has(character)) {
       // Found the end of a tag, so push the current tag.
       if (!utils.string.stringIsEmpty(currentTag)) {
         const parsedHedTag = new ParsedHedTag(
@@ -504,13 +506,162 @@ const formatHedTag = function (hedTag) {
 }
 
 /**
+ * Substitute certain illegal characters and report warnings when found.
+ */
+const substituteCharacters = function (hedString) {
+  const issues = []
+  const illegalCharacterMap = { '\0': ['ASCII NUL', ' '] }
+  const flaggedCharacters = /[^\w\d./$ :-]/g
+  const replaceFunction = function (match, offset) {
+    if (match in illegalCharacterMap) {
+      const [name, replacement] = illegalCharacterMap[match]
+      issues.push(
+        generateIssue('invalidCharacter', {
+          character: name,
+          index: offset,
+          string: hedString,
+        }),
+      )
+      return replacement
+    } else {
+      return match
+    }
+  }
+  const fixedString = hedString.replace(flaggedCharacters, replaceFunction)
+
+  return [fixedString, issues]
+}
+
+/**
+ * Check if group parentheses match. Pushes an issue if they don't match.
+ */
+const countTagGroupParentheses = function (hedString) {
+  const issues = []
+  const numberOfOpeningParentheses = utils.string.getCharacterCount(
+    hedString,
+    openingGroupCharacter,
+  )
+  const numberOfClosingParentheses = utils.string.getCharacterCount(
+    hedString,
+    closingGroupCharacter,
+  )
+  if (numberOfOpeningParentheses !== numberOfClosingParentheses) {
+    issues.push(
+      generateIssue('parentheses', {
+        opening: numberOfOpeningParentheses,
+        closing: numberOfClosingParentheses,
+      }),
+    )
+  }
+  return issues
+}
+
+/**
+ * Check if a comma is missing after an opening parenthesis.
+ */
+const isCommaMissingAfterClosingParenthesis = function (
+  lastNonEmptyCharacter,
+  currentCharacter,
+) {
+  return (
+    lastNonEmptyCharacter === closingGroupCharacter &&
+    !(
+      delimiters.has(currentCharacter) ||
+      currentCharacter === closingGroupCharacter
+    )
+  )
+}
+
+/**
+ * Check for delimiter issues in a HED string (e.g. missing commas adjacent to groups, extra commas or tildes).
+ */
+const findDelimiterIssuesInHedString = function (hedString) {
+  const issues = []
+  let lastNonEmptyValidCharacter = ''
+  let lastNonEmptyValidIndex = 0
+  let currentTag = ''
+  for (let i = 0; i < hedString.length; i++) {
+    const currentCharacter = hedString.charAt(i)
+    currentTag += currentCharacter
+    if (utils.string.stringIsEmpty(currentCharacter)) {
+      continue
+    }
+    if (delimiters.has(currentCharacter)) {
+      if (currentTag.trim() === currentCharacter) {
+        issues.push(
+          generateIssue('extraDelimiter', {
+            character: currentCharacter,
+            index: i,
+            string: hedString,
+          }),
+        )
+        currentTag = ''
+        continue
+      }
+      currentTag = ''
+    } else if (currentCharacter === openingGroupCharacter) {
+      if (currentTag.trim() === openingGroupCharacter) {
+        currentTag = ''
+      } else {
+        issues.push(generateIssue('invalidTag', { tag: currentTag }))
+      }
+    } else if (
+      isCommaMissingAfterClosingParenthesis(
+        lastNonEmptyValidCharacter,
+        currentCharacter,
+      )
+    ) {
+      issues.push(
+        generateIssue('commaMissing', {
+          tag: currentTag.slice(0, -1),
+        }),
+      )
+      break
+    }
+    lastNonEmptyValidCharacter = currentCharacter
+    lastNonEmptyValidIndex = i
+  }
+  if (delimiters.has(lastNonEmptyValidCharacter)) {
+    issues.push(
+      generateIssue('extraDelimiter', {
+        character: lastNonEmptyValidCharacter,
+        index: lastNonEmptyValidIndex,
+        string: hedString,
+      }),
+    )
+  }
+  return issues
+}
+
+/**
+ * Validate the full unparsed HED string.
+ *
+ * @param {string} hedString The unparsed HED string.
+ * @return {Issue[][]} String substitution issues and other issues.
+ */
+const validateFullUnparsedHedString = function (hedString) {
+  const [fixedHedString, substitutionIssues] = substituteCharacters(hedString)
+  const issues = [].concat(
+    countTagGroupParentheses(fixedHedString),
+    findDelimiterIssuesInHedString(fixedHedString),
+  )
+  return [substitutionIssues, issues]
+}
+
+/**
  * Parse a full HED string into a object of tag types.
  *
  * @param {string} hedString The full HED string to parse.
  * @param {Schemas} hedSchemas The collection of HED schemas.
- * @returns {[ParsedHedString, Issue[]]} The parsed HED tag data and list of issues.
+ * @returns {[ParsedHedString|null, Issue[]]} The parsed HED tag data and list of issues.
  */
 const parseHedString = function (hedString, hedSchemas) {
+  const [substitutionIssues, fullHedStringIssues] =
+    validateFullUnparsedHedString(hedString)
+  const fullStringIssues = fullHedStringIssues.concat(substitutionIssues)
+  if (fullHedStringIssues.length > 0) {
+    return [null, fullStringIssues]
+  }
   const parsedString = new ParsedHedString(hedString)
   const [hedTagList, splitIssues] = splitHedString(hedString, hedSchemas)
   parsedString.topLevelTags = findTopLevelTags(
@@ -526,8 +677,22 @@ const parseHedString = function (hedString, hedSchemas) {
   )
   formatHedTagsInList(parsedString.tags)
   formatHedTagsInList(parsedString.topLevelTags)
-  const issues = [].concat(splitIssues, tagGroupIssues)
-  return [parsedString, issues]
+  const parsingIssues = fullStringIssues.concat(splitIssues, tagGroupIssues)
+  return [parsedString, parsingIssues]
+}
+
+/**
+ * Parse a set of HED strings.
+ *
+ * @param {string[]} hedStrings A set of HED strings.
+ * @param {Schemas} hedSchemas The collection of HED schemas.
+ * @return {[ParsedHedString[], Issue[]]} The parsed HED strings and any issues found.
+ */
+const parseHedStrings = function (hedStrings, hedSchemas) {
+  const parsedHedStringsAndIssues = hedStrings.map((hedString) => {
+    return parseHedString(hedString, hedSchemas)
+  })
+  return zip(...parsedHedStringsAndIssues)
 }
 
 module.exports = {
@@ -539,4 +704,5 @@ module.exports = {
   splitHedString: splitHedString,
   formatHedTag: formatHedTag,
   parseHedString: parseHedString,
+  parseHedStrings: parseHedStrings,
 }
